// src/app/models/inventory.ts
/**
 * Descreve classe Ativo
 */
export class Inventory {
  _id: string;
  id: string;
  assetNum: number;
  subAssetNum: number;
  class: number;
  description: string;
  descriptionComp: string;
  inventoryNum: number;
  capitalizedOn: Date;
  costCenter: number;
  logPrinter: [{ message: string }];
}